backend/requirements.txt
###########################################################
# Web Framework
###########################################################
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv>=1.0.1
pyyaml==6.0.1

###########################################################
# ML & NLP Core
###########################################################
transformers>=4.41.0,<5.0.0
sentence-transformers>=2.3.0
numpy>=1.24.0,<2.0.0
pandas>=2.1.0
accelerate>=0.25.0
huggingface-hub>=0.20.0

###########################################################
# Torch + Torchaudio + Torchvision (호환 버전)
# ⚠️ 주의: TTS(Text-to-Speech)와 Deep Agent 모두 PyTorch 필요!
# ⚠️ 절대로 제거하지 마세요!
#
# Windows 환경 (권장):
#   pip install torch torchvision torchaudio
#
# Linux + AMD GPU (ROCm):
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/rocm6.0
#
# NVIDIA GPU (CUDA):
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
###########################################################
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

###########################################################
# Vector DB
###########################################################
chromadb>=0.4.18

###########################################################
# Date / Utils
###########################################################
python-dateutil>=2.8.0
packaging>=21.0
tqdm
loguru==0.7.2

###########################################################
# OpenAI / HTTP
###########################################################
openai>=1.55.0
httpx>=0.27.0

###########################################################
# STT / Audio I/O
###########################################################
sounddevice>=0.5.0
soundfile>=0.12.1
librosa==0.9.1
pydub==0.25.1
pyaudio==0.2.14
soxr>=0.3.7

###########################################################
# TTS & Text Processing (MeloTTS)
###########################################################
txtsplit
cached_path>=1.6.0
num2words>=0.5.13
unidic_lite==1.0.8
unidic==1.1.0
pykakasi==2.2.1
fugashi==1.3.0
g2p_en==2.1.0
anyascii==0.3.2
jamo==0.4.1
gruut[de,es,fr]==2.2.3
g2pkk>=0.1.2
python-mecab-ko>=1.3.0
eng_to_ipa==0.0.2
inflect==7.0.0
unidecode==1.3.7
pypinyin==0.50.0
cn2an==0.5.22
jieba==0.42.1
langid==1.1.6

###########################################################
# Whisper / Voice
###########################################################
faster-whisper>=1.0.0
resemblyzer>=0.1.1

###########################################################
# LangChain / LangGraph
###########################################################
langchain>=1.0.0,<2.0.0
langchain-openai>=1.0.0,<2.0.0
langgraph>=1.0.0,<2.0.0
langgraph-prebuilt>=1.0.4,<2.0.0
nest_asyncio>=1.5.8

# Authentication (Google OAuth + JWT)
###########################################################
PyJWT>=2.8.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0

###########################################################

# Excel
##########################################################
openpyxl>=3.0.0
pandas>=2.0.0
##########################################################

##########################################################
# Deep Agent Pipeline
diffusers>=0.25.0
tenacity>=8.2.0
pillow>=10.0.0

# Qwen 2.5 14B GGUF (Local Scenario Writer)
# llama-cpp-python==0.2.90
##########################################################